cm 21/08/2017 app restructuring

+++ machine_code/execution_master_interactive.sh
@@ -1,8 +1,6 @@
-#!/bin/bash
-
 #----------------------------------------------------------------------------#
 
-# Purpose:     Master Execution Script - Cron Job
+# Purpose:     Master Execution Script - Interactive Mode
 # Author:      CM
 # Date:        July 2017
 # Language:    Shell (.sh)
@@ -11,22 +9,8 @@
 
 # Settings
 #----------------------------------------------------------------------------#
-## read in saved settings
 printf "\n# Settings\n\n"
-
-# chmod a+rwx cron_user_setting
-while read -r line
-do
-    eval $line
-done < cron_user_setting
-# chmod a-rwx cron_user_setting && chmod o+rwx cron_user_setting 
-cd ${wd_path}
-
-printf "\n# Initialising\n\n"
-source code_base/machine_code/setting.sh 
-
-## update date
-export email_date=$(date -v -${length}d +"%d-%B-%Y") 
+source code_base/machine_code/user_setting.sh 
 
 #----------------------------------------------------------------------------#
 #                         Step-by-Step Tool Execution                        #
@@ -38,6 +22,13 @@ export email_date=$(date -v -${length}d +"%d-%B-%Y")
 #----------------------------------------------------------------------------#
 #----------------------------------------------------------------------------#
 
+# Status
+#----------------------------------------------------------------------------#
+$CD bubble --title "Email Analyser" \
+--text "Starting the Analysis" ‑‑no‑timeout \
+--background-top "F8F8F8" --background-bottom "F8F8F8" --border-color "F8F8F8" \
+--icon-file "${wd_path_helper}/icon/icon_temp.png"
+
 # Stage-a: Obtain emails
 #----------------------------------------------------------------------------#
 printf "\n# Obtaining Emails\n"
@@ -45,7 +36,6 @@ printf "\n# Obtaining Emails\n"
 # execute
 #---------------------------------------------------#
 cd ${wd_path_code}/stage_a
-echo "$(which php)"
 ${php_custom_path} -c ${php_custom_path_ini} -f "email_extract.php"
 
 # Stage-b (i): Process emails
@@ -57,8 +47,8 @@ printf "\n# Classifying Emails\n"
 cd ${wd_path_code}/stage_b
 
 ## classify
-${R_path} CMD BATCH --no-save "--args ${init_path} ${execution_id} ${data_path_train} \
-${data_path_raw_outbox} ${data_path_temp} ${wd_path_log} ${wd_path_model} ${lib_path}" classify.R \
+R CMD BATCH --no-save "--args ${init_path} ${execution_id} ${data_path_train} \
+${data_path_raw_outbox} ${data_path_temp} ${wd_path_log} ${wd_path_model}" classify.R \
 ${wd_path_log}/classify_${execution_id}.Rout
 
 ## delete output file
@@ -73,29 +63,13 @@ printf "\n# Generating Report\n"
 cd ${wd_path_code}/stage_b
 
 ## aggregate
-${R_path} CMD BATCH --no-save "--args ${init_path} ${execution_id} ${data_path_temp} \
-${wd_path_helper_email} ${wd_path_log} ${lib_path} ${email_address}" aggregate.R ${wd_path_log}/aggregate_${execution_id}.Rout
+R CMD BATCH --no-save "--args ${init_path} ${execution_id} ${data_path_temp} \
+${wd_path_helper_email} ${wd_path_log} ${email_address}" aggregate.R ${wd_path_log}/aggregate_${execution_id}.Rout
 
 ## delete output file
 [ -e .RData ] && rm .RData
 
-# Stage-c: Send email
-#----------------------------------------------------------------------------#
-printf "\n# Sending Email\n"
-
-# execute
-#---------------------------------------------------#
-cd ${wd_path_code}/stage_c
-
-## obtain file names
-email_report_file=$(ls ${data_path_temp}/*report_${execution_id}*)
-export email_report=`cat ${email_report_file}`
-export email_plot=$(ls ${data_path_temp}/*mean_plot_${execution_id}*)
-
-## aggregate
-${php_custom_path} -c ${php_custom_path_ini} -f  "send_email.php"
-
-# Stage-d: Clear 
+# Stage-b (iii): Clear 
 #----------------------------------------------------------------------------#
 printf "\n# Clearing Data\n"
 
@@ -103,18 +77,36 @@ printf "\n# Clearing Data\n"
 cd ${data_path_raw_outbox}
 rm email*txt
 
+## create output folder
+mkdir ${wd_path_output}/output_${current_date}_${execution_id}
+
 ## move output to location & delete all data
 cd ${data_path_temp}
+mv *mean_plot_${execution_id}* ${wd_path_output}/output_${current_date}_${execution_id}/how_am_I_doing_plot_${current_date}.pdf
+mv *report_${execution_id}* ${wd_path_output}/output_${current_date}_${execution_id}/how_am_I_doing_report_${current_date}.txt
 rm *${execution_id}*
 
 ## delete log files
 cd ${wd_path_log}
-rm *${execution_id}*
+cp classify_${execution_id}*txt ${wd_path_output}/output_${current_date}_${execution_id}/how_am_I_doing_log_${current_date}.txt
+if [ "${keep_log}" == "No" ]; then
+	cd ${wd_path_log}
+	rm *${execution_id}*
+else 
+	printf "\n# Log files kept - See ${wd_path_log}\n"
+fi
 
 # Stage-d: Output
 #----------------------------------------------------------------------------#
-printf "\n# Completed\n"
+printf "\n# Completed - See the report & graph at ${wd_path_output}/output_${current_date}_${execution_id}/\n"
 cd ${wd_path}
+
+# Status
+#----------------------------------------------------------------------------#
+$CD bubble --title "Email Analyser" \
+--text "Successfully Completed - (output_${current_date}_${execution_id})" ‑‑no‑timeout \
+--background-top "F8F8F8" --background-bottom "F8F8F8" --border-color "F8F8F8" \
+--icon-file "${wd_path_helper}/icon/icon_temp.png"
 
 #----------------------------------------------------------------------------#
 #                                    End                                     #
